ls_17 reorganized again
