+++ store/src/app/products/cart/cart.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { Product } from 'src/app/interfaces/products.interface';
-import { ProductsService } from '../../services/products.service';
+import { Product } from 'src/app/shared/interfaces/products.interface';
+import { ProductsService } from '../../shared/services/products.service';
 import { CartService } from './cart.service';
 
 @Component({
