+++ store/src/app/products/local-storage.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Product } from '../interfaces/products.interface';
+import { Product } from '../shared/interfaces/products.interface';
 
 @Injectable({
   providedIn: 'root'
