+++ store/src/app/products/products.module.ts
@@ -5,10 +5,10 @@ import { ProductCardComponent } from './product-card/product-card.component';
 import { ProductInfoComponent } from './product-info/product-info.component';
 import { ButtonComponent } from './button/button.component';
 import { FooterComponent } from './footer/footer.component';
-import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
+import { PageNotFoundComponent } from '../shared/components/page-not-found/page-not-found.component';
 import { RouterModule } from '@angular/router';
-import { HryvniaPipePipe } from '../pipes/hryvnia-pipe.pipe';
-import { ColorPriceDirective } from '../directives/color-price.directive';
+import { HryvniaPipePipe } from '../shared/pipes/hryvnia-pipe.pipe';
+import { ColorPriceDirective } from '../shared/directives/color-price.directive';
 import { ProductsComponent } from './products.component';
 import { ProductsListComponent } from './products-list/products-list.component';
 import { CartComponent } from './cart/cart.component';
